+++ src/Http/Controllers/CP/ShippingRatesController.php
@@ -3,18 +3,15 @@
 namespace Jonassiewertsen\StatamicButik\Http\Controllers\CP;
 
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Cache;
 use Jonassiewertsen\StatamicButik\Blueprints\ShippingRateBlueprint;
 use Jonassiewertsen\StatamicButik\Http\Controllers\CpController;
 use Jonassiewertsen\StatamicButik\Http\Models\ShippingRate;
-use Jonassiewertsen\StatamicButik\Http\Models\ShippingZone;
-use Statamic\CP\Column;
 
 class ShippingRatesController extends CpController
 {
     public function store(Request $request)
     {
-//        $this->authorize('store', ShippingRate::class); TODO: Add authorization
+        $this->authorize('store', ShippingRate::class);
 
         $blueprint = new ShippingRateBlueprint();
         $fields = $blueprint()->fields()->addValues($request->all());
@@ -25,7 +22,7 @@ class ShippingRatesController extends CpController
 
     public function update(Request $request, ShippingRate $shippingRate)
     {
-//        $this->authorize('update', $shippingProfile);
+        $this->authorize('update', $shippingRate);
 
         $blueprint = new ShippingRateBlueprint();
         $fields = $blueprint()->fields()->addValues($request->all());
@@ -36,7 +33,7 @@ class ShippingRatesController extends CpController
 
     public function destroy(ShippingRate $shippingRate)
     {
-//        $this->authorize('delete', $shippingProfile);
+        $this->authorize('delete', $shippingRate);
 
         $shippingRate->delete();
     }
